refactor: Restructure project into modular format
Restructured the project to support a more modular design, separating common
utilities and specific scrapers.

- Created new directories for common utilities and scrapers:
  src/{common,pdf,goodreads}
- Implemented common utilities:
  - Created src/common/file_utils.py for file-related functions
  - Created src/common/web_utils.py for web-related functions
- Moved relevant code to the common files and updated imports in pdf/scraper.py
  - Updated imports in pdf/scraper.py to use functions from common utilities
- Modified src/main.py to run the PDF scraper
  - Main script will be further updated to accommodate all different
    libraries in the future
